Bug #9194
Uses now ProfiledObjectId in the signature of several methods instead of using
two parameters (one for the object id and the other for the object type).

In the user notification, check the users and the groups in the recipients of
the notification can be notified: in the case of a notification about a
resource, it means the user or the group of users can access the resource.