Bug #9194
Fix the migration version of the module busCore.

Remove the system notification about an access right profile update or deletion.
This is to avoid listeners to be used with this event as long as the
authorization profiles managment isn't refactored to be atomic and correct.
The deletion of the subscriptions are now just done when a component instance or
a node is deleted by listening events about such deletion and not more about
the deletion of an access right profile. For doing, the ComponentInstManager
notifies now when a component instance is deleted; the previous explicit call of
unsubscription done by some application services are removed.

Move the method isGroupAuthorized(final String, final T) from both
NodeAccessControl and ComponentAccessControl to the interface AccessController
with a default implementation that throws an exception. The implementations
NodeAccessController and ComponentAccessController overrides the default
implementation.

AbstractResourceUserNotificationBuilder checks now also if the user or the group
can access a node when the notification is about a node as resource.