Bug #9194: adjusting some stuffs
Adjusting the algorithm of node verification rights against a group.
Indeed, it is not necessary to walk into the node path in order to get the parent on which the specific rights are set.
If a node has inherited rights but it is a child of an other one with specific rights, the rightsdependson data is set with the id of the parent with specific rights.

Adding the verification of node rights of group by checking also parent groups of verified group (Administration).

Group identifier and component instance identifier where flipped into a call of ComponentAccessControl.get().isGroupAuthorized(...) from isGroupCanAccess method of AbstractSubscriptionBean?!!

Avoiding to call several times isValid method of AbstractSubscrptionBean in a same context. Indeed this method is induicing some CPU and database processing.

Into notification builder, moving the access control processing of user and group before the part of treatment which determines if the build process has to stop.
Moving also the knowledge of aliases management into component user notification building implementations instead of trying to guess from the transversal one.
When the resource is neither a contribution, neither a node, verifying access control  first on component instance identifier if any before grant the user.