Feature #13179
Replace the use of view:comments by viewTag:displayComments.
Later is defined as a JSTL tag fragment instead of a taglib tag.
The view:comment is kept for instance but it is planned to be removed.
This is the first step in the rewrite of the comment widget to
separate the GUI build from the GUI behaviour.