Feature #13179
Implement a way for a user to go back to the page in which he has asked
to sign in by clicking on the connection button of the comment block of
a given contribution. The authentication process is delegated to the
auto-redirection mechanism in Silverpeas. By using this mechanism,
whatever the service performing the authentication (Silverpeas itself or
by SS0 service), the redirection to the page from which the user has
asked for signing in will be correctly treated. In fact, the
authentication process will be triggered by the auto-redirection
mechanism itself as the user behind the request isn't identified
(because it is hidden behind an anonymous navigation and hence session).

For doing, when the user asks to sign in from the comments block, he's
redirected to the GoToContribution servlet. This servlet is the entry
point in the auto-redirection mechanism. In order this way of
auto-redirection mechanism works, some adjustments was required in it:
- now, anonymous navigation isn't more bound to a single Silverpeas user
  session but, like any well-identified users, to a dedicated Silverpeas
  user session built upon the HTTP session.
- In order to ensure the user with an anonymous session can be
  redirected to the contribution page from which he comes, once
  authenticated, the redirection data are transfered from the anonymous
  session to the new user session (the anonymous session is closed).
- Auto-redirection servlet AutoRedirectServlet doesn't treat anymore an
  anonymous session like an undefined session. Redirection to the
  authentication process is invoked for anonymous users only if the
  component instance targeted by the redirection isn't accessible by
  them or if no component instance can be identified (this is the case
  with the redirection invoked through the GoToContribution servlet).
- Because SessionManagement#getSessionInfo(String) treats also the
  anonymous sessions (as it is already expected in some codes), it
  requires in some other codes the returned session info isn't an
  anonymous one.
- No SessionInfo is an abstract class providing a common implementation
  for more concrete user session. HTTPSessionInfo is the concrete
  Silverpeas user session built upon HttpSession.
- As the redirection mechanism for the contributions (GoToContribution)
  is based upon the ComponentInstanceRoutingMap API. Later has been
  improved in order to be more generic.

A message informing the user to be authenticated to post comments or to
view comments is added with the connection button in the comments block.