Feature #13179
Instead of dispatching on Login.jsp, using directly /Login path when the aim is to redirect the user to the login page.

Renaming Login.jsp into welcome.jsp in order to avoid confusions.

Managing a new parameter with AutoRedirectServlet and Permalinks servlet that allow to force a user to login.